requirements.txt
django-cors-headers
psycopg2
tensorflow
opencv-python
Pillow
django
djangorestframework

